handel: move contribution into handel.go
